[10643] Update the ACE part we use to 5.8.2

+++ dep/ACE_wrappers/ace/SPIPE_Stream.inl
@@ -1,6 +1,6 @@
 // -*- C++ -*-
 //
-// $Id: SPIPE_Stream.inl 80826 2008-03-04 14:51:23Z wotte $
+// $Id: SPIPE_Stream.inl 88707 2010-01-25 18:49:25Z johnnyw $
 
 #include "ace/OS_NS_sys_uio.h"
 #include "ace/OS_NS_errno.h"
@@ -111,7 +111,7 @@ ACE_SPIPE_Stream::send_handle (ACE_HANDLE handle) const
   DWORD procID;
   WSAPROTOCOL_INFO protInfo;
   ssize_t res = this->recv(&procID, sizeof(procID));
-  if (res != sizeof(procID))
+  if (res != static_cast <ssize_t> (sizeof(procID)))
   {
     if(res != -1)
         errno = ENXIO;
@@ -123,7 +123,7 @@ ACE_SPIPE_Stream::send_handle (ACE_HANDLE handle) const
       return -1;
   }
   res = this->send(&protInfo, sizeof(protInfo));
-  if (res != sizeof(protInfo))
+  if (res != static_cast <ssize_t> (sizeof(protInfo)))
   {
     if(res != -1)
         errno = ENXIO;
@@ -131,7 +131,7 @@ ACE_SPIPE_Stream::send_handle (ACE_HANDLE handle) const
   }
   // This is just for synchronization, we will ignore the data
   res = this->recv(&procID, sizeof(procID));
-  if (res != sizeof(procID))
+  if (res != static_cast <ssize_t> (sizeof(procID)))
   {
     if(res != -1)
         errno = ENXIO;
@@ -165,14 +165,14 @@ ACE_SPIPE_Stream::recv_handle (ACE_HANDLE &handle) const
   pid_t procID = ACE_OS::getpid();
   WSAPROTOCOL_INFO protInfo;
   ssize_t res = this->send(&procID, sizeof(procID));
-  if (res != sizeof(procID))
+  if (res != static_cast <ssize_t> (sizeof(procID)))
   {
     if(res != -1)
         errno = ENXIO;
     return -1;
   }
   res = this->recv(&protInfo, sizeof(protInfo));
-  if (res != sizeof(protInfo))
+  if (res != static_cast <ssize_t> (sizeof(protInfo)))
   {
     if(res != -1)
         errno = ENXIO;
@@ -187,7 +187,7 @@ ACE_SPIPE_Stream::recv_handle (ACE_HANDLE &handle) const
   // Since it does not matter what the data is, just send something to
   // synchronize the end of the exchange
   res = this->send(&procID, sizeof(procID));
-  if (res != sizeof(procID))
+  if (res != static_cast <ssize_t> (sizeof(procID)))
   {
     if(res != -1)
         errno = ENXIO;
